fix(coordinator): convert RunState/ClientState to u8, Model to blob for Pod safety
- Changed  and  (ClientState) fields from enums to
- Added type-safe accessor methods: get_run_state(), set_run_state(), get_client_state(), set_client_state()
- Changed  from Model enum to  (blob) with get_model()/set_model() accessors
- Added serde_with::Bytes for serde compatibility and ts-rs type annotation
- Resolves Undefined Behavior in unsafe Pod implementation for Solana zero-copy deserialization

+++ shared/coordinator/src/data_selection.rs
@@ -1,6 +1,6 @@
 use crate::{Committee, CommitteeSelection, Coordinator, Round};
 
-use psyche_core::{BatchId, ClosedInterval, NodeIdentity, deterministic_shuffle};
+use psyche_core::{deterministic_shuffle, BatchId, ClosedInterval, NodeIdentity};
 use std::{collections::BTreeMap, fmt};
 
 /// Assigns data batches to nodes based on committee roles.  
@@ -195,7 +195,7 @@ mod tests {
         let clients: Vec<_> = (0..num_nodes)
             .map(|i| Client {
                 id: TestNode(i as u64),
-                state: ClientState::Healthy,
+                state: ClientState::Healthy as u8,
                 exited_height: 0,
             })
             .collect();
